_archive/projects/fullpotential_ai/fullpotential_core/agents/services/i-proactive/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
redis>=5.0.1
crewai>=1.0.0
litellm>=1.0.0
mem0ai>=0.1.0
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.1
httpx>=0.25.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
psutil>=5.9.0
chromadb>=0.4.22
